Update patch set 1
Patch Set 1:

Reworked the functionality a bit after the comments. Jedi mind trick is now enabled when you reach Jedi Padawan. Each title after that increases the success chance for the mind trick by absolute 5%. Having the skill Force Powers Master is the only way to achieve 100% success rate (the skill gives the player mind trick that works on NPC's so I thought that was a good way to also give it this benefit).

Patch-set: 1
Label: Verified=0