Update patch set 2
Patch Set 2: Code-Review-1

Can you edit the commit message again so that it displays what the original fix was for? You can do this within gerrit by clicking on "Commit Message" and then clicking the clipboard looking icon in the top right of the file. 

What you should do in the future is use 'git commit --amend'. This will keep the old commit message and allow you to add on to it. As well as the original change ID. 

Also you should join the #opendev channel on irc  @ irc.swgemu.com

Patch-set: 2
Label: Code-Review=-1