Update patch set 1
Patch Set 1: Code-Review-2

Thank you for your contribution! We developers have had a pretty long discussion about how to best solve the issue with areas not being returned correctly from the QuadTree. The QuadTree is a central part of the server and heavily used so performance is very important as well in this part. Your change showed us that we need to take a new path for handling areas in the server. 

The QuadTree is a point tree and not really suited for handling areas. You will always risk missing areas that are large, even with your modifications. The QuadTree might not even do the inRange check for the area of interests since it has stored the area in another part of the tree. The storage is based on the center of the region. Our current plan now is to do some internal testing performance vice for a few different solutions. Based on that we will select the solution that we think will be best in the long run. We will also implement tests to verify that areas are working as expected, something that the current code is missing. 

A few general comments about your commit that you can keep in mind for all future commits that we hope you will submit: 

 - Set up you editor to remove trailing white spaces. 
 - Increase your knowledge about polymorphism, check the following page on how you can get different behavior from the same method by overriding it in sub classes: https://www.tutorialspoint.com/cplusplus/cpp_polymorphism.htm 
 - ManagedReference<> is only needed when you want to store a reference to an object in another object. You can use Reference<> for temporary objects. 

Do not hesitate to ask for questions or help in the dev channel on our discord server! You are welcome to ask me direct questions as well if you have any.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Code-Review=-2