Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Please update the already merged spawnPoint arrays with the new noDelay value. As it is now the code relies on Lua returning a non 1 value. It is also risky to omit values and rely on default values if we need to add even more values to the spawnPoint structs.

Patch-set: 2
Label: Code-Review=-1