Update patch set 1
Patch Set 1:

Pulled this locally and did a little bit of research in the forum archive, then discussed it with Miztah. Looks like if you remove the templateType, gameObjectType and the machineType it will show in the crafting tool properly under Furniture and list on the bazaar properly as well.

Patch-set: 1
CC: Gerrit User 4976 <4976@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>