Update patch set 1
Patch Set 1: Code-Review-2

This issue is something we have discussed previously and do not believe this is the proper method to address this issue. We do understand that it is a problem where BH's are attacking Jedi during PvP and it makes it near impossible for the Jedi to stay in the fight. This is where ideas like anonymous BH missions came from.

There will be a time in the future where we can look to find a method to better handle the BH/Jedi and group pvp dynamic that doesn't have possible negative outcomes for others where outside heals could come into play.

Patch-set: 1
Reviewer: Gerrit User 4976 <4976@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Code-Review=-2