Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> This issue is something we have discussed previously and do not believe this is the proper method to address this issue. We do understand that it is a problem where BH's are attacking Jedi during PvP and it makes it near impossible for the Jedi to stay in the fight. This is where ideas like anonymous BH missions came from.
> 
> There will be a time in the future where we can look to find a method to better handle the BH/Jedi and group pvp dynamic that doesn't have possible negative outcomes for others where outside heals could come into play.

The anonymous BH missions while also a positive change merely lowers the likelihood of this situation occurring without mitigating the impact of the event.  If this is an issue that has been recognized by staff, why would we not be looking at both mitigating the likelihood and impact of this issue?  The proposed change directly mitigates the immediate impact of a BH in group GCW scenarios, and is a very small change.  Possibly the smallest change that could be made to address this issue other than completely removing the BH TEF.  I understand fully that 14.1 is the focus right now, but with such a minor change behind a config flag, why is this not at least given a shot, tested with Finalizer and removed if deemed to actually have a negative impact. Even if this change is temporary until a more robust solution is determined.

For clarification exactly who did you mean by "others" in the below statement, as by the submitted changes, it can only be the BH in this case:

"group pvp dynamic that doesn't have possible negative outcomes for others where outside heals could come into play"

The Jedi already has negative outcomes from taking damage from multiple sources (GCW & BH) while the BH does not.  The negative outcomes from the Jedi now being effectively removed from GCW PVP when a BH comes has a larger impact on the Jedi and all the players engaged in the GCW PVP.  By negating the BH TEF GCW PvP for healing, this does not necessarily negatively impact the BH, it only mitigates their massive advantage that already exists, and removes their impact on an entire group of players.  If anything with these changes, the opposing faction of the Jedi with a BH TEF will still gain an advantage in the fight, which will still benefit the BH, this only creates a more level playing field.  I hope this review is reconsidered internally.

Patch-set: 1
Reviewer: Gerrit User 5378 <5378@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>