[fixed] use defaultInput in SuiInputBox and set it for renaming - Mantis #4197
Not quite sure about why the isFilterBox() condition was there.  For
me it looks like it does make more sens if the condition would have
been negated.  Removed it all together for now defaultInput is set
always as well as maxInputSize.  In my testing I could not find
anything that broke due to the change.

Also added proper calls to set the defaultInput to VendorManager and
StructureManager - ContainerImplementation.cpp did already have the
defaultInput setup correctly.

Change-Id: I1bc3099e8c292a70ea655993d667bf2dd60ca6b6