Fix issues preventing server shutdown
When shutting down, the server currently hangs after clearing the final
ground zone. This is caused by new zoneCleared flag which hides the one
from the base class. Once this issue is resolved, the next problem is a
segmentation fault while clearing the space zones. This is due to typos
in the octree logic, which causes it to repeatedly clear the northwest
node from the first layer rather than any nodes from the second layer.

Change-Id: I12e132a3f26189f88af03d46bd0176c6fc2e89e6