Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 1: Code-Review-1
> 
> (2 comments)

From what I can tell, the jtlTravelPoints property is only ever used to resolve the arrival point when a client sends a LaunchIntoSpace command. Since the arrivalPointName included in this command comes from planetTravelPoints, the names in jtlTravelPoints must match.

The strings mentioned in the comments refer to conversation options which are used for the SpaceStationScreenPlay. This screenplay's travel points are stored in a separate file and are not involved in this change.

Patch-set: 2
Reviewer: Gerrit User 6076 <6076@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>